requirements.txt
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
curl_cffi==0.13.0
frozendict==2.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
llvmlite==0.44.0
multitasking==0.0.12
numba==0.61.2
numpy==2.2.6
pandas==3.0.2
pandas-ta==0.4.71b0
peewee==4.0.4
platformdirs==4.9.4
protobuf==7.34.1
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2026.1.post1
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
